Exercício 14-Pessoa UPDATE_02

+++ Códigos/Exercícios/14-Pessoa/Professor.java
@@ -1,6 +1,6 @@
 public class Professor extends Pessoa02{
     private  String especialidade;
-    private double salario;
+    private float salario;
 
     public void receberAum(float s){
         setSalario(getSalario()+s);
@@ -16,11 +16,11 @@ public class Professor extends Pessoa02{
         this.especialidade = especialidade;
     }
 
-    public double getSalario() {
+    public float getSalario() {
         return salario;
     }
 
-    public void setSalario(double salario) {
+    public void setSalario(float salario) {
         this.salario = salario;
     }
